no option to check every configured master during a slave check for a zone
the default value is "no" and keeps the old behavior (choosing random master
for SOA-check) except that the master used for the SOA check is also used for
the possible zone transfer (old behavior uses always the first master which
breaks zone transfers if this master is down although other masters
are still working).

when "yes", every master is used for SOA check and the master with the
highest serial is used for zone transfer.